use local-data-api:0.6.12 & env vars for setup script override
local-data-api:0.5.8 was returning empty list records for queries that return list values. This resulted in the rds-data boto3 clients to error saying list values were not supported. The newer version of local-data-api fixes this issue but requires using the secret/cluster ARNs that were used for the local metadb setup for the install script within the terraform module. To pass the local metadb arns, the metadb setup script looks for the env vars and if not then uses the tf module's ARN's.